Updated Enemy AI
Player now can damage the enemy.
Enemy will start chasing the player when the player hits it; But only if the enemy is able to see the player (An raycast is thrown when the player damages the enemy. If the raycast hits the player, the enemy will start chasing him; If not, the enemy will not move)